@@ -9,9 +9,9 @@
   "scripts": {
     "build": "babel src/ -d dist/ --copy-files",
     "start": "node dist/",
-    "dev": "nodemon --exec babel-node src/",
-    "dev:debug": "nodemon --exec babel-node --inspect=0.0.0.0:9229 src/index.js",
-    "lint": "eslint src --config .eslintrc.js --fix",
+    "dev": "nodemon --exec babel-node src/ -e js,graphql",
+    "dev:debug": "nodemon --exec babel-node --inspect=0.0.0.0:9229 src/index.js -e js,graphql",
+    "lint": "eslint src --config .eslintrc.js",
     "test": "nyc  --reporter=text-lcov yarn test:jest",
     "test:cypress": "run-p --race test:before:*",
     "test:before:server": "cross-env GRAPHQL_URI=http://localhost:4123 GRAPHQL_PORT=4123 babel-node src/ 2> /dev/null",
@@ -36,19 +36,19 @@
     ]
   },
   "dependencies": {
-    "activitystrea.ms": "^2.1.3",
-    "apollo-cache-inmemory": "~1.4.3",
-    "apollo-client": "~2.4.13",
-    "apollo-link-context": "^1.0.14",
-    "apollo-link-http": "~1.5.11",
-    "apollo-server": "~2.4.2",
+    "activitystrea.ms": "~2.1.3",
+    "apollo-cache-inmemory": "~1.5.1",
+    "apollo-client": "~2.5.1",
+    "apollo-link-context": "~1.0.14",
+    "apollo-link-http": "~1.5.12",
+    "apollo-server": "~2.4.8",
     "bcryptjs": "~2.4.3",
     "cheerio": "~1.0.0-rc.2",
-    "cors": "^2.8.5",
+    "cors": "~2.8.5",
     "cross-env": "~5.2.0",
     "date-fns": "2.0.0-alpha.26",
     "dotenv": "~6.2.0",
-    "express": "^4.16.4",
+    "express": "~4.16.4",
     "faker": "~4.1.0",
     "graphql": "~14.1.1",
     "graphql-custom-directives": "~0.2.14",
@@ -57,39 +57,37 @@
     "graphql-shield": "~5.3.0",
     "graphql-tag": "~2.10.1",
     "graphql-yoga": "~1.17.4",
-    "helmet": "^3.15.1",
+    "helmet": "~3.15.1",
     "jsonwebtoken": "~8.5.0",
     "linkifyjs": "~2.1.8",
     "lodash": "~4.17.11",
     "ms": "~2.1.1",
-    "neo4j-driver": "~1.7.2",
-    "neo4j-graphql-js": "~2.3.1",
+    "neo4j-driver": "~1.7.3",
+    "neo4j-graphql-js": "~2.4.1",
     "node-fetch": "~2.3.0",
     "npm-run-all": "~4.1.5",
-    "passport": "~0.4.0",
-    "passport-jwt": "~4.0.0",
-    "request": "^2.88.0",
+    "request": "~2.88.0",
     "sanitize-html": "~1.20.0",
     "slug": "~1.0.0",
     "trunc-html": "~1.1.2",
-    "uuid": "^3.3.2",
+    "uuid": "~3.3.2",
     "wait-on": "~3.2.0"
   },
   "devDependencies": {
     "@babel/cli": "~7.2.3",
-    "@babel/core": "~7.3.3",
+    "@babel/core": "~7.3.4",
     "@babel/node": "~7.2.2",
     "@babel/plugin-proposal-throw-expressions": "^7.2.0",
-    "@babel/preset-env": "~7.3.1",
+    "@babel/preset-env": "~7.3.4",
     "@babel/register": "~7.0.0",
-    "apollo-server-testing": "~2.4.2",
+    "apollo-server-testing": "~2.4.8",
     "babel-core": "~7.0.0-0",
     "babel-eslint": "~10.0.1",
-    "babel-jest": "~24.1.0",
+    "babel-jest": "~24.3.1",
     "chai": "~4.2.0",
-    "cucumber": "^5.1.0",
-    "debug": "^4.1.1",
-    "eslint": "~5.13.0",
+    "cucumber": "~5.1.0",
+    "debug": "~4.1.1",
+    "eslint": "~5.15.1",
     "eslint-config-standard": "~12.0.0",
     "eslint-plugin-import": "~2.16.0",
     "eslint-plugin-jest": "~22.3.0",
@@ -97,7 +95,7 @@
     "eslint-plugin-promise": "~4.0.1",
     "eslint-plugin-standard": "~4.0.0",
     "graphql-request": "~1.8.2",
-    "jest": "~24.1.0",
+    "jest": "~24.3.1",
     "nodemon": "~1.18.10",
     "nyc": "~13.3.0",
     "supertest": "~3.4.2"
