@@ -7,12 +7,10 @@ export const host = 'http://127.0.0.1:4123'
 export async function login ({ email, password }) {
   const mutation = `
       mutation {
-        login(email:"${email}", password:"${password}"){
-          token
-        }
+        login(email:"${email}", password:"${password}")
       }`
   const response = await request(host, mutation)
   return {
-    authorization: `Bearer ${response.login.token}`
+    authorization: `Bearer ${response.login}`
   }
 }
